Limit adding blogs only for authorized users

+++ part4/bloglist/app.js
@@ -1,8 +1,10 @@
 const express = require("express");
 const mongoose = require("mongoose");
+const jwt = require("jsonwebtoken");
 const { info, error } = require("./utils/logger");
 const { mongodb_uri } = require("./utils/config");
 const Blog = require("./models/blog");
+const User = require("./models/user");
 
 mongoose.connect(mongodb_uri, { useNewUrlParser: true, useUnifiedTopology: true })
     .then(() => info("Connected to MongoDB"))
@@ -10,6 +12,17 @@ mongoose.connect(mongodb_uri, { useNewUrlParser: true, useUnifiedTopology: true 
 mongoose.set("useCreateIndex", true);
 mongoose.set("useFindAndModify", true);
 const router = express.Router();
+
+const identifyUser = request => {
+    const authHeader = request.get("authorization");
+    if (!(authHeader && authHeader.toLowerCase().startsWith("bearer")))
+        return null
+    const token = authHeader.substring(7);
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!(token && decodedToken.id))
+        return null;
+    return await User.findById(decodedToken.id);
+}
 
 router.get("/", async (_, response, next) => {
     try {
@@ -21,9 +34,14 @@ router.get("/", async (_, response, next) => {
 });
 
 router.post("/", async (request, response, next) => {
+    const user = identifyUser(request);
+    if (!user)
+        return response.status(401).json({ detail: "Token missing or invalid" });
+
     try {
         if (!("likes" in request.body))
             request.body.likes = 0;
+        request.body.author = user.username;
         const blog = new Blog(request.body);
         const newBlog = await blog.save();
         response.status(201).json(newBlog).end();
@@ -33,6 +51,10 @@ router.post("/", async (request, response, next) => {
 });
 
 router.patch("/:id", async (request, response, next) => {
+    const user = identifyUser(request);
+    if (!user)
+        return response.status(401).json({ detail: "Token missing or invalid" });
+
     try {
         const body = request.body;
         const res = await Blog.findOneAndUpdate({ id: request.param.id }, body, { new: true });
@@ -43,6 +65,10 @@ router.patch("/:id", async (request, response, next) => {
 });
 
 router.delete("/:id", async (request, response, next) => {
+    const user = identifyUser(request);
+    if (!user)
+        return response.status(401).json({ detail: "Token missing or invalid" });
+
     try {
         Blog.findOneAndDelete({ id: request.params.id });
         response.status(200).end();
